requirements.txt
Pillow==9.0.1
matplotlib==3.5.1
pyperclip==1.8.2